+++ mmpretrain/models/multimodal/chinese_clip/__init__.py
@@ -1,6 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .bert import BertModelCN
-from .bert_tokenizer import FullTokenizer
 from .chinese_clip import ChineseCLIP, ModifiedResNet
 
-__all__ = ['ChineseCLIP', 'ModifiedResNet', 'FullTokenizer', 'BertModelCN']
+__all__ = ['ChineseCLIP', 'ModifiedResNet', 'BertModelCN']
