+++ mmpretrain/models/utils/tokenizer.py
@@ -1,9 +1,13 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import os
+import collections
 
 from transformers import (AutoTokenizer, BartTokenizer, BertTokenizer,
-                          BertTokenizerFast, LlamaTokenizer)
+                          BertTokenizerFast, LlamaTokenizer, BasicTokenizer,
+                          WordpieceTokenizer)
+from mmengine.fileio import list_from_file
 
+from mmpretrain.registry import TOKENIZER
 from .huggingface import register_hf_tokenizer
 
 register_hf_tokenizer(AutoTokenizer)
@@ -110,3 +114,74 @@ class OFATokenizer(BartTokenizer):
         tokenizer.bin_offset = length + 8192
         tokenizer.num_bins = num_bins
         return tokenizer
+
+
+@TOKENIZER.register_module()
+class FullTokenizer(BertTokenizer):
+    """Runs end-to-end tokenziation."""
+
+    def __init__(self, vocab_file, do_lower_case=True):
+        self.vocab = self.load_vocab(vocab_file)
+        self.inv_vocab = {v: k for k, v in self.vocab.items()}
+        self.basic_tokenizer = BasicTokenizer(do_lower_case=do_lower_case)
+        self.wordpiece_tokenizer = WordpieceTokenizer(
+            vocab=self.vocab, unk_token='[UNK]', max_input_chars_per_word=200)
+
+    def load_vocab(self, vocab_file):
+        """Loads a vocabulary file into a dictionary."""
+        vocab = collections.OrderedDict()
+        index = 0
+        vocab_list = list_from_file(vocab_file)
+        for token in vocab_list:
+            if not token:
+                break
+            token = token.strip()
+            vocab[token] = index
+            index += 1
+        return vocab
+
+    def tokenize(self, text):
+        split_tokens = []
+        for token in self.basic_tokenizer.tokenize(text):
+            for sub_token in self.wordpiece_tokenizer.tokenize(token):
+                split_tokens.append(sub_token)
+
+        return split_tokens
+
+    def convert_by_vocab(self, vocab, items):
+        """Converts a sequence of [tokens|ids] using the vocab."""
+        output = []
+        for item in items:
+            output.append(vocab[item])
+        return output
+
+    def convert_tokens_to_ids(self, tokens):
+        return self.convert_by_vocab(self.vocab, tokens)
+
+    def convert_ids_to_tokens(self, ids):
+        return self.convert_by_vocab(self.inv_vocab, ids)
+
+    @staticmethod
+    def convert_tokens_to_string(tokens, clean_up_tokenization_spaces=True):
+        """Converts a sequence of tokens (string) in a single string."""
+
+        def clean_up_tokenization(out_string):
+            """Clean up a list of simple English tokenization artifacts like
+            spaces before punctuations and abbreviated forms."""
+            out_string = (
+                out_string.replace(' .', '.').replace(' ?', '?').replace(
+                    ' !', '!').replace(' ,', ',').replace(" ' ", "'").replace(
+                        " n't", "n't").replace(" 'm", "'m").replace(
+                            " 's", "'s").replace(" 've",
+                                                 "'ve").replace(" 're", "'re"))
+            return out_string
+
+        text = ' '.join(tokens).replace(' ##', '').strip()
+        if clean_up_tokenization_spaces:
+            clean_text = clean_up_tokenization(text)
+            return clean_text
+        else:
+            return text
+
+    def vocab_size(self):
+        return len(self.vocab)
